Changed id in Product entity to be of type Long instead of int

+++ src/main/java/com/waldronprojects/bookstore/dao/ProductDao.java
@@ -6,12 +6,12 @@ import com.waldronprojects.bookstore.entity.Product;
 
 public interface ProductDao {
 
-	public Product getProduct(int id);
+	public Product getProduct(Long id);
 	
 	public List<Product> getProducts();
 
 	public void saveProduct(Product product);
 	
-	public void deleteProduct(int id);
+	public void deleteProduct(Long id);
 
 }
